Move defined_mouse_buttons() to layout-util
Rename to layout_handle_user_defined_mouse_buttons().
