+++ src/view-file/view-file-list.cc
@@ -33,6 +33,7 @@
 #include "filedata.h"
 #include "intl.h"
 #include "layout-image.h"
+#include "layout-util.h"
 #include "layout.h"
 #include "main-defines.h"
 #include "metadata.h"
@@ -40,7 +41,6 @@
 #include "options.h"
 #include "ui-fileops.h"
 #include "ui-menu.h"
-#include "ui-misc.h"
 #include "ui-tree-edit.h"
 #include "utilops.h"
 #include "view-file.h"
@@ -537,7 +537,7 @@ gboolean vflist_release_cb(ViewFile *vf, GtkWidget *widget, GdkEventButton *beve
 	GtkTreeIter iter;
 	FileData *fd = nullptr;
 
-	if (defined_mouse_buttons(bevent, vf->layout))
+	if (layout_handle_user_defined_mouse_buttons(vf->layout, bevent))
 		{
 		return TRUE;
 		}
